Do not download the file if pyang is already installed

+++ install.sh
@@ -10,20 +10,25 @@ cd $DIR
 GIT_BRANCH="$(git rev-parse --abbrev-ref HEAD)"
 GIT_COMMIT_HASH="$(git log -1 --format=%h)"
 
-sudo pip install -r requirements.txt
-sudo rm -rf /tmp/pyang-swagger
-mkdir /tmp/pyang-swagger
+if pyang -v | grep -q '1.7.2'; then
+  	echo "Pyang version 1.7.2 is already installed"
+  	sudo cp $(pwd)/pyang/plugins/swagger.py /usr/local/lib/python2.7/dist-packages/pyang-1.7.2-py2.7.egg/pyang/plugins/swagger.py
+else
+	sudo pip install -r requirements.txt
+	sudo rm -rf /tmp/pyang-swagger
+	mkdir /tmp/pyang-swagger
 
-wget https://github.com/mbj4668/pyang/archive/pyang-1.7.2.tar.gz -O /tmp/pyang-swagger/pyang.tar.gz
-mkdir /tmp/pyang-swagger/pyang
-tar -zxf /tmp/pyang-swagger/pyang.tar.gz -C /tmp/pyang-swagger/pyang/ --strip-components=1
+	wget https://github.com/mbj4668/pyang/archive/pyang-1.7.2.tar.gz -O /tmp/pyang-swagger/pyang.tar.gz
+	mkdir /tmp/pyang-swagger/pyang
+	tar -zxf /tmp/pyang-swagger/pyang.tar.gz -C /tmp/pyang-swagger/pyang/ --strip-components=1
 
-#git clone https://github.com/sebymiano/pyang-swagger.git
-cp $(pwd)/pyang/plugins/swagger.py /tmp/pyang-swagger/pyang/pyang/plugins/
+	#git clone https://github.com/sebymiano/pyang-swagger.git
+	cp $(pwd)/pyang/plugins/swagger.py /tmp/pyang-swagger/pyang/pyang/plugins/
 
-cd /tmp/pyang-swagger/pyang
-sudo pip install -r requirements.txt
-sudo python setup.py install
+	cd /tmp/pyang-swagger/pyang
+	sudo pip install -r requirements.txt
+	sudo python setup.py install
+fi
 
 CONFIG_PATH=$HOME/.config/iovnet/
 FILE_NAME=pyang-swagger.yaml
@@ -38,5 +43,7 @@ cat > $CONFIG_PATH$FILE_NAME << EOF
 git-info: ${GIT_BRANCH}/${GIT_COMMIT_HASH}
 install-date: ${DATE}
 EOF
+
+echo "PYANG-SWAGGER installed successfully"
 
 cd $_pwd
